Refactor: @TagMaxSize 리팩토링 (#34)
