+++ src/main/java/kr/teammanagers/tag/exception/validator/TagMaxSize.java
@@ -1,4 +1,4 @@
-package kr.teammanagers.tag.exception;
+package kr.teammanagers.tag.exception.validator;
 
 import jakarta.validation.Constraint;
 import jakarta.validation.Payload;
@@ -8,15 +8,17 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
+import static kr.teammanagers.tag.constant.TagErrorConstant.MAX_TAG_SIZE;
+
 @Constraint(validatedBy = TagMaxSizeValidator.class)
 @Target({ElementType.METHOD, ElementType.FIELD, ElementType.ANNOTATION_TYPE, ElementType.PARAMETER})
 @Retention(RetentionPolicy.RUNTIME)
 public @interface TagMaxSize {
-    String message() default "태그는 최대 {size}개까지만 설정 가능합니다.";
+    String message();
 
     Class<?>[] groups() default {};
 
     Class<? extends Payload>[] payload() default {};
 
-    int size() default 3;
+    int size() default MAX_TAG_SIZE;
 }
